Implement manager for repository histories in partition
In the previous commit, we implemented a history that keeps track
of changes in a given repository. Each partition may contain more
than one repository though, so we need a type to manage multiple
histories.

Histories need to be created when the repository receives its first
writes. Transactions need to be evicted from the histories once they
are no longer needed for conflict checking to release the memory
associated with them.

This commit implements a history manager that keeps track of the
reference histories in a partition. TransactionManager can then use
this type to get the history for a correct repository, and apply
changes to it. The type also tracks which history a transaction at
a given LSN modified and routes the evictions to the correct history.
This way transaction manager can simply call evict on the correct LSN,
and the history manager handles the complexity of dropping the now
unneeded state.