storagemgr: Make RequireRepositories print failing relative path
The RequireRepositories test utility adds an additional "require.Failf"
statement if it detects a failure in any of the asserting repositories.
This statement has an undesirable side-effect: it swallows the
stacktrace of the data race. The stacktrace is printed out after the
test finishes. "require.Failf" prints failed tests itself and then calls
"t.FailNow()" to exit the tests instantly.

This commit updates the statement to use "tb.Log" instead.