Abort writes that ran concurrently with a repository deletion
We're currently allowing writes to commit even if the target repository
was concurrently deleted. The writes are generally conflict checked though
using the actual state of the repository on the disk, so the changes are
still verified to be correct if the repository also happened to be
concurrently recreated. We're about to start performing conflict checks
in-memory based on the change history. This behavior causes problems with
verifying the logical reference changes in-memory. If a repository is
deleted, we'd have to record every reference as deleted for the conflict
checks to work.

As it is not a common use case to delete and recreate the repository while
performing writes to it, just abort transactions if the repository was
deleted after they took their read snapshot. This allows us to keep the
upcoming conflict checks simpler.

To avoid expanding transaction manager further, we implement the conflict
check in a new type. For now, it just performs the deletion conflict check.
We'll later expand it to cover more of the conflict checks, and generalize
the deletion check to a file level conflict check.