Trigger race in tests via API
UserMergeBranch tests are testing that the RPC fails if the target
branch is concurrently updated. It does so by modifying the repository
directly on the disk. As this update does not go through the API and
TransactionManager, it won't soon be noticed when transaction manager
starts verifying reference conflicts in-memory.

Trigger the race in the test via the API.