migration: Introduce migrations
Repositories need a way to perform custom one-time operations that get
committed to the WAL. These operations are plain go functions that, when
executed, modify the specified transaction is some defined manner.

To faciliate this, introduce the `migration.migration` type. When
executed, the migration performs the configured function and writes a
migration key to indicate the migration was been performed.