Return expected error from UserMergeBranch
UserMergeBranch returns an error detail when reference update fails
due to a concurrent update on the reference. The error returned to
the RPC handler was recently changed in d3e14502 but the error handling
wasn't updated to convert it to the expected error detail. Convert it
now to retain backwards compatibility in the API.