Expand reference tree type
Reference tree type is a helper used for recording reference changes
that should be logged into WAL. We're about to start using the type
in a more general manner to record the existing state of references
in the repository. As we apply the updates we record, we also need
to keep the in-memory representation of the current reference db
structure up to date. Expand the type with functionality to remove
and insert arbitrary nodes.

The internal representation had to be slightly changed and the
hierarchy now starts from the repository root, allowing the 'refs'
directory to be represented in the hierarchy as well. The improved
handling here also caught buggy test assertions. When HEAD is updated,
we shouldn't log twice the creation of 'refs' directory.

The type is placed into its own package as it has grown quite a lot.