repository: Remove WithSkipRepositoryExistenceCheck option
WithSkipRepositoryExistenceCheck option was added to repository
creating RPCs as the repository is not supposed to exist yet.
It was also added to RemoveRepository RPC during that refactoring
but this looks like a mistake because repoutil.Remove will return
error if the repository doesn't exists anyway. So we should check
the existence of the repository and return earlier.