Record reference tree changes in memory
We're removing 'git update-ref' from the critical section of
transaction manager as it is slow. To make this possible, we need
to record the file changes reference transactions perform outside
of the critical section. Recording the changes on each invocation
of the reference transaction hook is the easiest way to do this as
every reference update invokes the hook. The problem there is that
we don't have a good way to read the pre-image of the repository
before any changes are made to later figure out from the post-image
of the update which directories and files were created or removed.
This is because the prepared stage of the hook already acquires locks
and may create directories where it needs to create the locks.

This commit moves the existing reference recording logic into a
separate type that keeps track of the on-disk state of the reference
db in memory. The pre-image of the reference db is recorded before
any reference changes are made. The pre-image can then be compared to
the post-image after the reference transaction has been applied, and
the in-memory pre-image is modified afterwards to ensure it is accurate
for a subsequent reference transaction.

In practice, this allow us to record the reference database's pre-image
when the transaction starts, and record the changes on each invocation of
the 'committed' stage of the reference transaction hook. As the pre-image
is kept up to date with the changes, subsequent 'committed' stage invocations
of the reference transaction hook can also record their changes.

To keep the changes smaller, we'll for now just plug the recorder where
we currently verify the references in the critical section. Later on,
we can record the changes while the transaction is executing to stop
running update-ref in the critical section.