Record packed-refs changes in ReferenceRecorder
ReferenceRecorder is currently only recording changes inside the
'refs' directory and to the HEAD file. To fully record all changes
performed, we should also record the changes to the 'packed-refs'
file. This commit adds support to the recorder for logging packed-refs
changes. It records the pre-image of the packed-refs file by linking
it to a temporary directory. After all changes are performed, it
compares the inode of the pre-image file to the inode of the post-image
file. If they are different, it logs the file.

The packed-refs file is only logged once all of the changes are done.
This avoids logging it multiple times if there are multiple deletions.

We also integrate this for a good measure in TransactionManager. It's
a no-op in its current location though as TransactionManager is applying
the updated packed-refs file to the repository before running updates
to avoid rewriting the file in the critical section. We'll later move
this whole recording procedure outside the critical section.

GetInode function is moved to the WAL package. We keep it exported as
the TransactionManager is still handling the packed-refs staging and
uses the function.