storage: Extract some directory state helpers to common testhelper
In the storagemgr package, we have some testing utilities for asserting
local directory state and setting log entries. Those utilities are
located in the test files of the package. In some following commits,
we'll need them in tests of other packages. This commit moves them to
a sub-package of common testhelper package. They could not be moved to
testhelper package directly due to circular dependency issue.