Take in git.ReferenceUpdates to RecordReferenceUpdates
RecordReferenceUpdates is taking in the protobuf representation of
reference changes. This is a remain from the time when we'd eventually
feed those to update-ref there. Since we no longer do that, take
in the domain type representing a reference transaction to simplify
and make the function easier to call from other locations without
converting the updates to protobuf.