Migrate object repacking to physical logging protocol
Housekeeping is the only piece of functionality that does not use
the physical logging protocol but has its own custom application
logic. We're looking to migrate it to the common logging approach
for simplicity, and to record the file system operations housekeeping
performs so we can later conflict check the operations on a file level.

This commit removes the custom application logic related to object
repacking, and instead records the file system operations to do when
performing the repacking. On verification, the operations are replayed
on to the snapshot of the target repository to verify they apply.
Otherwise the verification part is unchanged. The application relies
on the existing logic to apply file system changes.

As the custom message in the WAL entry schema is now unnecessary, it
is also removed.