Migrate commit graph rewriting to physical logging protocol
We're migrating housekeeping to use the physical logging protocol
like every other operation. This commit migrates the commit-graph
rewriting part of it. We follow the same pattern as the existing
custom application logic by removing all of the commit-graphs and
afterwards committing the result of the rewrite.

Commit graphs had no verification logic, so the no-op function that
existed was removed. The application is now using the same logic
to apply the file system changes as other operations.

As the custom message for commit graph rewriting is no longer needed,
it is removed.