Implement file system level conflict checks
We're working towards generalizing the conflict checks in order to
remove the usage of Git from the TransactionManager's critical
section as using Git there has a major negative impact on performance.
Git is currently used there to replay the recorded reference changes
on top of the current state of the repository. If conflicts arise,
the transcation is aborted.

To do this without Git, we've recently implemented logic to verify
concurrent reference changes against each other. This allows us to
say whether or not the logical reference changes conflict. In addition
to ensuring the logical references do not conflict, we need to ensure
the file system changes needed to represent those reference changes do
not conflict. This commit implements the logic to conflict check
concurrent file system changes against each other.

The conflict checks assume that transaction are correct in isolation,
and they only attempt to commit changes that are valid in their own
snapshot. This assumption allows us to do these conflict checks in
memory as we don't need to validate the changes against the base state.
If the change staged are correct in isolation, the only possible source
of conflicts are changes committed concurrently by other transactions.

Each transaction's file system changes are recorded in a tree and
transactions are verified one-by-one. The first transaction to commit
sees an empty tree, and the changes thus won't conflict with anything.
Subsequent transactions will see the earlier transaction's state in the
tree. The changes are then conflict checked with the following logic:

- The changes must produce a valid directory tree. For example,
  non-empty directories can't be removed. This could occur if another
  transaction concurrently added files to the directory.
- Concurrent modifications to the same node are not allowed, for example
  if the first transaction replaces a file, second transaction that was
  concurrently executing is not allowed to remove the file as it couldn't
  have possibly seen the changes made by the first transaction. This is
  achieved by tracking the LSN of the transaction that last modified a
  node and comparing it to the LSN of the read snapshot of a subsequent
  writer.
- There's also support for conflict checking reads. This can be used to
  verify that a file read during transaction did not change concurrently.
  While we generally wouldn't record all of the writes a transaction does,
  this can be used to prevent write skew. In practice, we'll use this to
  prevent the alternates file of a repository being concurrently modified
  when we base our changes on the current state of the alternate.

The conflict checks are possibly a bit too strict at the moment as we for
example reject writes into directories that were concurrently created.
This could produce conflicts if two transcations concurrently create a
directory like 'refs/heads/my-directory' and write 'ref-1' and 'ref-2'
into it. The other directory creation would be safe to drop. We'll start
with overly strict checks and relax them where we deem it safe to do so.

The History keeps track of all changes made until they are expicitly
evicted from the history. TransactionManager will evict changes of an LSN
from the History when it no longer needs to keep them around for conflict
checking. This is the case when all active writes have the changes already
in their read snapshot.

We don't yet integrate the conflict checks into TransactionManager but
will do so in later changes.