backup: Let backup work directly with underlying LogManager
In prior commits, the log management functionality is extracted out of
TransactionManager. The backup functionality followed this extraction.
The new LogManager is now responsible for tracking all indices as well
as consumer positions. However, the backup package still contacts the
TransactionManager for acknowledgement. This commit breaks up this
relationship. The backup now works directly with the new LogManager
instead of detouring through the TransactionManager.