backup: Add an integration test that uses real WAL log manager
In the tests of the internal/backup package, we mock the log manager
extensively. It was because it's not trivial to set a Transaction The
manager inside this package. Using mock comes with a cost. We noticed
that the type assertion in the implementation never passed outside of
tests.

Now the log management is detached from Transaction Manager. It makes
more sense to write an integration test so that the log archiver uses a
real WAL log manager. This test eliminates the aforementioned problem.