migration: Initialize migration key on repository creation
The migration key is used to determine if there are any outstanding
migrations required for a specific repository. When a repository gets
created it should have a key created in the same transaction as the
repository is created in. When created, new repositories are expected to
be up-to-date. Consequently, the value of the migration key should be
the highest configured migration ID. Since there are currently no
migrations configured, the key value will be initialized to zero.