ci: Increase tmpfs by 1GB
We've started seeing some flake in the Praefect tests when FIPS is enabled
(https://gitlab.com/gitlab-org/gitaly/-/jobs/8375481755) which appears to be
caused by running out of disk space in the tmpfs RAM disk volume we create.

Increase the size by 1GB to mitigate these test failures. It's possible that
we might not be cleaning up temporary files properly.