storage: Store referenced log entry in memory
TransactionManager keeps a list of committed but still referenced by
other transactions. At verification step, TransactionManager walks
through the list, load respective log entries back to the memory. This
loading step is redundant and repetitive. A committed entry is possibly
referenced by multiple transactions. When the TransactionManager
iterates through the list, an entry is loaded and unmarshaled multiple
times, leading to extra memory occupation. The life span of an item of
the list is likely to be short-term. Thus, it makes sense to keep the
log entry in memory.