Migrate reference packing to physical WAL protocol
Reference packing is the only operation that is not modeled by
general file system operations but has custom application logic.
This adds complexity and should be refactored to be in line with
all other transactions. More importantly this makes it difficult to
conflict check housekeeping operations against other transactions as
we don't have an exhaustive list of files and directories the
houskeeping transactions perform. This makes it difficult to generalize
the conflict checks to work on the level of file and directory writes.

This commit migrates the reference packing operation to log the file
system operations it performs, and removes the custom application logic
associated with it. The packed-refs file and the possible reference locks
are already staged before entering the critical section. The loose
reference removals are logged in the critical section as we don't
have a good way yet to handle conflicting writes. We'll later improve
this when the generalized conflict checks land and log all of the
operations before entering the critical section.