Check if node has already been evicted
EvictLSN evicts all paths from the tree that were last updated by
a given LSN. It may not be possible to evict some nodes if they still
have children that were updated by later LSN. Such nodes will be
cleaned up later as part of evicting those later LSNs. When evicting
a node, we walk up the tree and evict also all nodes that we couldn't
previously evict but now can. Whether a node is eligible to be evicted
as part of this walk is determined by whether it has children and whether
it is less than or equal to the evicted LSN.

This can cause a nil pointer panic if for example we're evicting a
transaction that updated `refs/parent` and `refs/parent/child`. If
we evict 'refs/parent/child' first and evict `refs/parent` while walking
towards the root, we'll panic when later attempting to access `refs/parent`
to evict it.

Fix the issue by checking whether the node has already been evicted.