Snapshot only existing parent directories
Snapshots have so far always included the parent directory hierarchy
of a repository even if the parent directories didn't exist. We're
about to add conflict checking for file system operations. We need
to know whether the parent directories exist or not to decide whether
the parent directories should be created as a part of the transaction
or not. This commit creates the parent directories in the snapshot
only if they actually exist in the storage.