Introduce a recording file system implementation
We're about to introduce general file system level conflict checks.
Using these checks, we'll move all of the operations to be recorded
outside of the critical section of the transaction manager and only
perform conflict checking there. We need an easy way to record the
file system operations a transaction is performing. This commit
introduces a type that wraps common file system operations so we can
record them as they are performed. For now, only Mkdir and MkdirAll
are supported as we'll need them soon.

We'll later expose this type through the Transaction interface. This
will enable transactions to stage arbitrary file system operations.