Log parent directory creation only when needed
TransactionManager is currently always logging repository parent
directory creation when a repository is created. This relies on the
fact that we ignore already exists errors on application as these
directories could already be created by another partition, or through
an interrupted application attempt.

We're about to start conflict checking file system operations. These
non-conditional parent directory creations will start conflicting when
we do this as we'd always log a creation even if the directories exist.
This commit changes this and records the parent directory creation only
if they were created during the transaction to avoid these conflicts.

As the repository's final state is still logged in the TransactionManager,
we also have to for now record the parent directory creation in Begin.
We'll later move the entire repository creation logging out from the
TransactionManager.

We remove the previous helper that was used to log repository creations
as it is no longer needed.