Integrate file level conflict checks
This commit plugs in file level conflict checks into transaction
manager. File level conflict checks are done at the end, so not much
changes yet as the existing conflict checks catch most of the conflicts
still. After we have file level conflict checks in place, we'll start
removing the various custom conflict checks and rely on these more
general ones.

The file level conflict checks work on the file system operations the
transaction has logged. We go through each of them and check whether
they still apply in light of all other concurrent operations that were
done. The history of file system changes is kept in memory to avoid
needing a repository snapshot where to apply the operations.

TransactionManager synthesizes read operations on each file and
directory that is being written/removed. While typically write-write
operations are not conflicts alone, we don't have a general way to
track reads the transaction has performed. We thus assume that every
write was preceded by a read on the file to produce a conflict.

For now, we don't conflict checks loose reference writes to allow
loose reference creations to commit after a pack-refs task that removed
the loose reference file. This is safe to do as reference changes have
already been verified for logical conflict before this step. The history
still verifies that the resulting file system state is valid, ie. the
file doesn't exist, and the parent directories exist.

Similarly conflicts of tables.list file are not checked. It conflicts on
every single reference write as all transactions need to update it with
the new reftable file. The conflicts are ignored as the transaction
manager is resolving the conflicts before performing the file conflict
checks.

We don't in general resolve conflicts yet. If transactions do concurrent
modifications on files or directories, they'll be aborted. This is safe
but may lead to aborting transactions that would be safe to commit.