Reset node's write LSN after a write
We're currently not resetting a node's writeLSN after a transaction
writes to it. This means that when a transaction reads a node that
it just wrote to, the checker considers it a conflict.

Reset the writeLSN on an update to reflect the fact that the
transaction is aware of the node, thus preventing a read-after-write
conflict due to the transaction's own operations.