Reset node's write LSN after a write
We're currently not resetting a node's write LSN after the
updated it. This leads to Read() calls conflicting with the write
even if the transaction performed it after deciding to ignore the
conflict. Reset the LSN on an update so the transaction's own writes
do not conflict with later read checks it performs.