storage: Remove redundant stale file cleaning up function
removeStaleWALFiles is not unnecessary these days as the log entry is
self-contained in the directory. Previously the manifest was in the
database with the pack files only in the directory. In fact this entire
function is likely unnecessary as log appending is an atomic rename, and
the pruning old entries could just happen in the usual code paths.