storage: Allow pruning log entries used by snapshots
We used to take the head of committed entries list into account when
determining which entries to be pruned. In a prior commit, we keep
referenced log entry state in the memory. Thus, there's no need to
prevent log from being pruned even if there are transactions with older
read snapshots around.

Recently, the TransactionManager has utilized a low watermark to limit
which log entries should be pruned. This low watermark takes the head of
committedEntries into account. Now, a log entry is pruned while it's
still referenced by other transactions. Thus, we cannot use the low
watermark anymore. The conflict manager now depends solely on the
committed list to evict an LSN. Unfortunately, `cleanCommittedEntry` is
called by some other goroutines. Thus, we need to use mutex to prevent
concurrent calls to `Prepare()` and `Evict()`.

This commit also set the initial consumer position to oldestLSN instead
of zero.