Record reference changes while the transaction runs
TransactionManager is currently recording the logical reference
changes during the transaction. The recorded reference transactions
are then replayed at commit time using `git update-ref` on an up to
date snapshot of the target repository. While this is a simple way
to verify the reference changes, it is extremely slow as it requires
repository snapshot, and using a Git process adds overhead. As the
reference files are recorded during the critical section, we also
need to fsync them during the critical section. This leads to multiple
fsyncs if a transaction has multiple reference changes. At minimum,
we still need four fsyncs even with just a single reference change
as we need to fsync the reference itself, the manifest file with
the instructions where to place it, and the WAL entry's directory
and finally move of the WAL entry to the WAL.

This commit takes a step towards removing this performance bottleneck.
Instead of replaying the reference changes at commit time, we'll
record them during the transaction. When the transaction begins, we
instantiate a reference recorder which records the current shape of
the refs directory and the packed-refs. When UpdateReferences is
called from the reference-transaction hook's committed phase, we
check what has changed and record the file system operations performed.

The recorded reference changes are conflict checked by the already
existing logical reference conflict checks. The file changes necessary
to perform the reference changes are conflict checked by the file level
conflict checks. Transaction is aborted if either the logical or the
physical changes conflict. We don't generally resolve conflicts yet so
transactions that make logically compatible changes may conflict. In
particular, reference writes that run concurrently with a pack-refs
operations and commit after it are rejected by thec conflict checks.
This can be improved later.

As update-ref is no longer needed, we remove its usage from the critical
section with files backend. Reftables still use it, and will need to
be migrated separately.

This change will eventually allow us to remove the repository
snapshotting from the critical section as we no longer need it for
update-ref. Some other operations still use it, so we leave it in place
for now. As we are now able to prepare the reference changes outside of
the critical section, we can also fsync them completely outside of the
critical section. This will allow for parallelizing fsyncs of multiple
transaction's WAL entries which will improve performance significantly.
For now, we still leave the fsyncing in the critical section in place
as we have some other operations still staging state there. We'll later
move them also out, and get down to a single fsync in the critical
section.