Perform default branch update before recording it
TransactionManager's test suite is recording the default branch
update before the update is actually performed. This didn't matter
before as TransactionManager recorded the reference changes to log
by replaying them, and the order was correct there. We are about to
start recording the file system changes associated with reference
updates during a transaction. We'll do this in the reference
transaction hook with a call to UpdateReferences. By the time
UpdateReferences is called the actual reference updates would have
been already peformed.

Change the ordering in the test to match what would happen in reality
so the default branch update tests don't start failing when we do this.