Write references to the snapshot in BenchmarkTransactionManager
BenchmarkTransactionManager is currently just calling UpdateReferences
to stage the reference updates. This will soon no longer work as we're
about to move the reference recording to happen when the updates are
performed in the snapshot as opposed to replaying them in the
transaction manager. Prepare for this by actually writing the references
into the snapshot in BenchmarkTransactionManager.