Append directory deletions at the end of operations list
Repository deletions are recorded in the critical section. The operations
were previously appended to the head of the operations list. This avoided
some issues that could occur if other operations were staged in the same
transaction. In reality, we should not have any operations that delete
the repository and stage other state while doing so. Simplify and just
append the operations to the end of the list. It's anyway the responsibility
of the transaction to produce a correct sequence of operations.