Fail if other operations are staged with repository deletions
We have a test in transaction manager asserting that a repository
deletion containing a reference modification succeeds. While this
may have asserted the behavior previously, it's not a realistic
scenario as our repository deleting transactions do not stage
reference updates. Abort repository deletions that contain other
updates in them to simplify the file level conflict checks were
about to introduce.

Repository deletions are staged in the critical section of the
transaction manager as we previously lacked generalized file level
conflict checks. Staging the deletions in the critical section meant
that we stage them using the latest state of the repository and that
they would be guaranteed to not conflict. The general file level
conflict checks will move allow for staging repository deletions
outside of the critical section based on the snapshot state. When
that is done, it becomes irrelevant as well whether reference updates
can be staged with a repository deletions as producing a correct
sequence of operations is no longer the responsibility of transaction
manager.