partition: Backup KV state of the partition
We implemented BackupPartition to backup the entire partition state
but the KV state of the partition is stored in an embedded DB shared
between all partitions, so it's not part of the file system snapshot.
With this change, we will include the partition KV in the same archive.

New proto type added for the KV entries as it provides more efficient
encoding for byte slices compared to default go encoding.