Clean temporary state of transactions in the background
TransactionManager is currently cleaning synchronously the temporary
state of a transaction it has processed after it finishes with it.
As the manager is single threaded, this increases commit latencies
as transactions have to wait until transaction manager has finished
cleaning up their temporary state. Worse, this adds the clean up
latency to every single transaction waiting in the queue to commit.
If the queue is long, this can add significant latency.

Median snapshot time of `gitlab-org/gitlab` observed on staging is
around 20ms but there are also outlier repositories that take a lot
longer. The snapshot clean up takes roughly the same amount of time.
This is doubled as we also need to clean up the snapshot we take in
the critical section to verify the transaction. Overall, this can
account for the majority of the commit latency.

This commit moves transaction clean up to a pool of background
workers. This allows transaction manager to keep processing further
transactions while the clean up happens in the background and thus
reducing latency significantly. We limit the number of background
workers available to exert backpressure on writes if we can't clean
up fast enough.

If the transaction was not admitted, the client goroutine is still
responsible for cleaning them up. This avoids saturating the pool
of background workers the manager uses.