Don't dereference symbolic references
TransactionManager contains some tests asserting that symbolic
references can be removed. The test suite however is not running
with the `--no-deref` to delete the symbolic reference itself
instead of targeting the reference it points to. This has worked
so far because the `update-ref` command used to replay the reference
changes in TransactionManager uses the flag. We're about to remove
the reference replaying and rely on the recorded reference changes
from the snapshot. The tests will fail once we do this as the changes
are not done in the snapshot as expected. Add the missing flag to the
test helper as well.

These tests are mostly for sanity checking. In reality, Gitaly does
not support writing symbolic references other than HEAD.