test: Fix FetchBundle tests that assert on Praefect votes
b886db48c6 (refs: don't invoke reference-transaction hook for reflogs, 2024-11-14)
was recently merged into the "next" branch of Git. This started causing
failures in the FetchBundle tests that assert on the content of Praefect
votes. This is because the upstream change starts omitting the HEAD
reference from reference transaction hooks.

Adjust the test to make two assertions. The first tests the old
behaviour and is allowed to fail, at which point we require the new
behaviour to be correct. A TODO is left to remove this once Gitaly
exclusively tests against Git versions that include the new behaviour.