migration: Update the migration id in the manager
The migration manager runs all the migrations to be executed in a loop.
Each migration is incharge of updating the repo's KV with the migration
ID once successful, to indicate that the migration corresponding to
that ID has been run successfully.

While this works, it makes more sense that the manager updates the ID
after each migration. This is for two reasons:
  - The manager is incharge of ensuring the order of migrations to be
  run and which to skip. In this sense, storing the last migration id
  should also be the responsibility of the owner.
  - Combing the migration core function with the update binds us from
  adding features like dry-run, wherein we want to run the core
  migration logic, but avoid the KV update.

In this sense, move the KV update responsibility to the manager.