storage: Move snapshot package to storage/storagemgr/partition/snapshot
The snapshot package creates file system snapshots from a given
directory hierarchy. This package is solely used by TransactionManager
to manage snapshots to be used by incoming transactions. The package is
located at `storage/storage/storagemgr/snapshot`. This commit moves the
package to the `partition` sub-package.

That way it would be better visible in the package structure as well
that it is a partition-level component. We've been moving restructuring
the package layer, and as part of that, we've tried to make the
layers/hierarchy better visible in the package structure. This is one of
the stragglers still that haven't been migrated.