packfile: Teach ReadIndex to use git command factory
The `packfile.ReadIndex` function currently fails the CI pipeline when
used in tests, as seen in
https://gitlab.com/gitlab-org/gitaly/-/jobs/8468733084.
This is because Gitaly prohibits reliance on the Git binary
from `$PATH`.

To address this issue, this commit updates `packfile.ReadIndex` to
utilize the Git command factory.
This change ensures compliance with Gitaly's Git binary enforcement,
allowing tests to pass without violating this constraint.