packfile: Move packfile.List to the git package and rename it
We aim to use the Git command factory in `packfile.ReadIndex`
because tests using `packfile.ReadIndex` currently fail the CI
pipeline, as seen in
https://gitlab.com/gitlab-org/gitaly/-/jobs/8468733084. This failure
occurs because we prohibit using the Git binary from `$PATH`.

However, the Git command factory depends on `packfile.List`,
which introduces a circular dependency if we use the Git command
factory in the `packfile` package.

To resolve this circular dependency, we move `packfile.List` to the
`git` package. Additionally, we rename `List` to `ListPackfiles` to
make its purpose more explicit.