Decouple fsrecorder from wal package
The wal package contains helpers to build a WAL entry. It currently
also hosts helpers to record various file system operations into
WAL. The fsrecorder.FS type is exposed through the transaction
interface to record file system operations. We want to move the file
system recording helpers there to expose them on the transaction.
To facilitate this, decouple the package by putting the WAL entry
builder behind an interface. This way we can test the logic without
having to assert details of the WAL entry.