Implement methods for recording file system operations
Transactions may need to commit arbitrary file changes. This commit
implements methods to do so on the fsrecorder.FS type and exposes
them through the storage.FS interface.

We add methods for recording file, hard link and directory creations,
and a method to remove directory entries. In addition, transactions
can record the reads they've performed. While the other operations are
used to persist changes, the reads are used for conflict checking when
the transaction is being committed. If a transaction's read overlaps
with another transactions write, there's a conflict.

We expose methods to record the fundamental operations only to keep
the surface level of the API small. The recording also doesn't change
the state in the snapshot which allows for recording operations done
outside of Gitaly by recording for example the changes done by Git.