Move file system recording helpers to storage package
We have various helpers for recording different file system operations
into a transaction. They've been historically located in the wal
package. This was fine as they were only used internally in the
transaction manager. Now that we have a general interface to record
file system operations in a transaction, we're going to move much of
this recording to happen in the snapshot during the transaction. The
recording will move out from the transaction manager into the handlers
responsible for the transactions.

To prepare for this, move the recording helpers out from the wal
package to the storage package where they can be used from. To facilitate
the move, the functions start recording their operations through the
public interface of the transaction.