Rename methods on wal.Entry for consistency
wal.Entry's methods are inconsistently named. Harmonize the names
to match across the methods to imperative actions we want to log.
This way the key value methods match the file system methods.
Mentioning recording separately is not necessary as the type already
makes the context clear; this is for building a WAL entry.